Bump Rust toolchain to nightly-2022-11-03

+++ src/rust/src/ios/api/call_manager_interface.rs
@@ -87,7 +87,7 @@ impl AppByteSlice {
         if self.bytes.is_null() {
             return None;
         }
-        Some(unsafe { slice::from_raw_parts(self.bytes, self.len as usize) })
+        Some(unsafe { slice::from_raw_parts(self.bytes, self.len) })
     }
 }
